Move list_ammo() and select_ammo()
